create custom timer hook
