+++ src/utils/getFormattedTime.ts
@@ -4,7 +4,12 @@ const getFormattedTime = (milliseconds: number) => {
   const hours = Math.floor((totalSeconds % (3600 * 24)) / 3600);
   const minutes = Math.floor((totalSeconds % 3600) / 60);
   const seconds = totalSeconds % 60;
-  return { days, hours, minutes, seconds };
+  return {
+    days: days < 10 ? "0" + days : days,
+    hours: hours < 10 ? "0" + hours : hours,
+    minutes: minutes < 10 ? "0" + minutes : minutes,
+    seconds: seconds < 10 ? "0" + seconds : seconds,
+  };
 };
 
 export default getFormattedTime;
